typescriptからjsに変更
